+++ R/Class_tskrrHeterogenousImpute.R
@@ -1,9 +1,10 @@
 #' Class tskrrHeterogenousImpute
 #'
 #' The class \code{tskrrHeterogenousImpute} is a subclass of the
-#' class \code{\link[xnet:tskrrHeterogenous-class]{tskrrHeterogenous}}
+#' class \code{\link[xnet:tskrrHeterogenous-class]{tskrrHeterogenous}} and
+#' \code{\link[xnet:tskrrImpute-class]{tskrrImpute}}
 #' specifically for heterogenous networks with imputed values. It is
-#' the result of the function \code{\link{impute_tskrr}} or the function \code{\link{impute_loo}}.
+#' the result of the function \code{\link{impute_tskrr}}.
 #'
 #' @slot y the matrix with responses
 #' @slot k the eigen decomposition of the kernel matrix for the rows
@@ -21,26 +22,13 @@
 #' are integers indicating the row resp column number.
 #' @slot imputeid a vector with integer values indicating which of
 #' the values in \code{y} are imputed
-#' @slot imputemethod a character value with either \code{"predictions"} or
-#' "loo" indicating whether the predictions or the leave-one-out
-#' crossvalidation values were used in the imputation.
-#' @slot loofun if imputemethod is "loo", this slot contains the loo
-#' function used during imputation.
 #' @slot niter an integer value gving the number of iterations used
 #' @slot tol a numeric value with the tolerance used
 #'
-#' @include Class_tskrrHeterogenous.R
+#' @include Class_tskrrHeterogenous.R Class_tskrrImpute.R
 #' @rdname tskrrHeterogenousImpute-class
 #' @aliases tskrrHeterogenousImpute
 #' @exportClass tskrrHeterogenousImpute
 setClass("tskrrHeterogenousImpute",
-         contains = "tskrrHeterogenous",
-         slots = c(imputeid = "integer",
-                   imputemethod = "character",
-                   loofun = "function",
-                   niter = "integer",
-                   tol = "numeric"),
-         prototype = list(
-           imputemethod = "",
-           loofun = function(){}
-         ))
+         contains = c("tskrrImpute", "tskrrHeterogenous")
+         )