+++ R/Class_tskrrHomogenousImpute.R
@@ -1,10 +1,10 @@
 #' Class tskrrHomogenousImpute
 #'
 #' The class \code{tskrrHomogenousImpute} is a subclass of the
-#' class \code{\link[xnet:tskrrHomogenous-class]{tskrrHomogenous}}
+#' class \code{\link[xnet:tskrrHomogenous-class]{tskrrHomogenous}} and
+#' \code{\link[xnet:tskrrImpute-class]{tskrrImpute}}
 #' specifically for homogenous networks with imputed values. It is
-#' the result of the function \code{\link{impute_tskrr}} or
-#' the function \code{\link{impute_loo}}.
+#' the result of the function \code{\link{impute_tskrr}}.
 #'
 #' @slot y the matrix with responses
 #' @slot k the eigen decomposition of the kernel matrix for the rows
@@ -23,27 +23,13 @@
 #' are integers indicating the row resp column number.
 #' @slot imputeid a vector with integer values indicating which of
 #' the values in \code{y} are imputed
-#' @slot imputemethod a character value with either \code{"predictions"} or
-#' "loo" indicating whether the predictions or the leave-one-out
-#' crossvalidation values were used in the imputation.
-#' @slot loofun if imputemethod is "loo", this slot contains the loo
-#' function used during imputation.
 #' @slot niter an integer value gving the number of iterations used
 #' @slot tol a numeric value with the tolerance used
 #'
-#' @include Class_tskrrHomogenous.R
+#' @include Class_tskrrHomogenous.R Class_tskrrImpute.R
 #' @rdname tskrrHomogenousImpute-class
 #' @aliases tskrrHomogenousImpute
 #' @exportClass tskrrHomogenousImpute
 setClass("tskrrHomogenousImpute",
-         contains = "tskrrHomogenous",
-         slots = c(imputeid = "integer",
-                   imputemethod = "character",
-                   loofun = "function",
-                   niter = "integer",
-                   tol = "numeric"),
-         prototype = list(
-           imputeid = integer(0),
-           imputemethod = "",
-           loofun = function(){}
-         ))
+         contains = c("tskrrImpute", "tskrrHomogenous")
+         )