+++ R/Class_tskrrTune.R
@@ -64,36 +64,20 @@ setValidity("tskrrTune", validTskrrTune)
 setMethod("show",
           "tskrrTune",
           function(object){
+
+            # HEADER
+
             ishomog <- is_homogenous(object)
-            type <- ifelse(ishomog,"Homogenous","Heterogenous")
+            type <- ifelse(ishomog,"homogenous","heterogenous")
             tl   <- ifelse(ishomog,"----------","------------")
-            cat(paste(type,"tuned two-step kernel ridge regression"),
-                paste(tl,"--------------------------------------",sep="-"),
+            cat(paste("Tuned",type,"two-step kernel ridge regression"),
+                  paste("-----",tl,"--------------------------------",sep="-"),
                 sep = "\n")
-            dims <- paste(dim(object@y), collapse = " x ")
-            cat("Dimensions:", dims,"\n")
-            cat("Lambda:\n")
-            print(lambda(object))
-
-            labs <- labels(object)
-            if(ishomog)
-              cat("\nLabels:")
-            else
-              cat("\nRow Labels:")
-
-            str(labs$k, give.length = FALSE, give.head = FALSE,
-                width = getOption("width") - 11)
-            if(!ishomog){
-              cat("Col Labels:")
-              str(labs$g, give.length = FALSE, give.head = FALSE,
-                  width = getOption("width") - 11)
-            }
+            .show_tskrr(object, ishomog)
 
             # Information on tuning
             excl <- object@exclusion
             if(object@replaceby0) excl <- paste(excl,"(values replaced by 0)")
-
-
 
             if(identical(object@loss_function, loss_mse))
               loss_name <- "Mean Squared Error (loss_mse)"
@@ -103,9 +87,11 @@ setMethod("show",
               loss_name <- "custom function by user"
 
             cat("\nTuning information:\n")
-            cat("  exclusion:",object@exclusion,"\n")
-            cat("  loss value:", object@best_loss,"\n")
-            cat("  loss function:", loss_name,"\n")
+            cat("-------------------\n")
+            cat("exclusion setting:",object@exclusion,"\n")
+            cat("loss value:", object@best_loss,"\n")
+            cat("loss function:", loss_name,"\n")
+
           })