+++ R/get_loo_fun.R
@@ -20,9 +20,10 @@
 #' interaction (i.e. a single cell). Therefor you do not have an argument
 #' \code{exclusion} for that method.
 #'
-#' For the class \code{tskrrTune}, not specifying \code{exclusion} or
-#' \code{replaceby0} returns the used loo function. If you specify either
-#' of them, it will use the method for the appropriate model and return
+#' For the classes \code{tskrrTune} and \code{tskrrImpute},
+#' not specifying \code{exclusion} or \code{replaceby0} returns the used
+#' loo function. If you specify either of them,
+#' it will use the method for the appropriate model and return
 #' a new loo function.
 #'
 #' @inheritParams loo